Dialogue Box Update 2.0: Movement
The box will now pop up after clicked by the log in button. It's set with an exmaple message and message header now but functions have been made to change these messages.
The continue button on the dialogue box sends the box back to its original unseen position.
Just have to connect this with the conditions of signing up/logging in and we'll have a fully functioning well designed dialogue box.