DIALOGUE BOX CATCH 22 FIXED MFFFFF
There was a damn await function being used in the template of the AWS code in the guide I was instructed to use. That's what was holding up all of the code coming after the call out to the userpool, that was the single called function screwing this entire thing up.
As it turns out, the guy who set up the tutorial shouldn't have been using the ConfigureWait function when sending out to the user pool. As I've found out, that should only be used under extremely specific circumstances.
Whoever made that AWS unity tutorial guide needs to get fired.